[Feat] 공격범위 초기설정

+++ Assets/Scripts/OJH/Data/HeroData/HeroData.cs
@@ -11,8 +11,8 @@ public class HeroData : ScriptableObject
     [SerializeField] private int _baseDamage;
     public int BaseDamage { get { return _baseDamage; } set { _baseDamage = value; } }
 
-    [SerializeField] private int _attackRange;
-    public int AttackRange { get { return _attackRange; } set { _attackRange = value; } }
+    [SerializeField] private float _attackRange;
+    public float AttackRange { get { return _attackRange; } set { _attackRange = value; } }
 
     [SerializeField] private float _attackDelay;
     public float AttackDelay { get { return _attackDelay; } set { _attackDelay = value; } }
